[GR-54582] Merged JNIFeature with JNIAutomaticFeature.
Other small simplifications.

(cherry picked from commit a77aa8de46af0dce8af67a24359bc1d4ef099d5b)
