+++ substratevm/src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/jni/JNIFeature.java
@@ -27,7 +27,15 @@ package com.oracle.svm.hosted.jni;
 import java.util.Arrays;
 import java.util.List;
 
+import org.graalvm.nativeimage.ImageSingletons;
+import org.graalvm.nativeimage.Platforms;
 import org.graalvm.nativeimage.hosted.Feature;
+import org.graalvm.nativeimage.impl.InternalPlatform;
+
+import com.oracle.svm.core.SubstrateOptions;
+import com.oracle.svm.core.feature.AutomaticallyRegisteredFeature;
+import com.oracle.svm.core.feature.InternalFeature;
+import com.oracle.svm.core.jni.JNIGeneratedMethodSupport;
 
 /**
  * Support for the Java Native Interface (JNI). Read more in JNI.md in the project's root directory.
@@ -35,10 +43,26 @@ import org.graalvm.nativeimage.hosted.Feature;
  * @see <a href="http://docs.oracle.com/javase/8/docs/technotes/guides/jni/">Java Native Interface
  *      Specification</a>
  */
-public class JNIFeature implements Feature {
+@AutomaticallyRegisteredFeature
+@Platforms(InternalPlatform.NATIVE_ONLY.class)
+public class JNIFeature implements InternalFeature {
+    @Override
+    public boolean isInConfiguration(IsInConfigurationAccess access) {
+        return SubstrateOptions.JNI.getValue();
+    }
 
     @Override
     public List<Class<? extends Feature>> getRequiredFeatures() {
         return Arrays.asList(JNIFunctionTablesFeature.class, JNICallWrapperFeature.class, JNILibraryLoadFeature.class);
+    }
+
+    @Override
+    public void duringSetup(DuringSetupAccess access) {
+        if (!ImageSingletons.contains(JNIGeneratedMethodSupport.class)) {
+            ImageSingletons.add(JNIGeneratedMethodSupport.class, new JNIGeneratedMethodSupport());
+        }
+        if (!ImageSingletons.contains(JNIJavaCallWrapperMethod.Factory.class)) {
+            ImageSingletons.add(JNIJavaCallWrapperMethod.Factory.class, new JNIJavaCallWrapperMethod.Factory());
+        }
     }
 }
